refactor: Refactor code to comply to basic principles
Fixes #65

Refactors code to comply with basic principles.

* **Separation of Concerns**:
  - Moves `AnimationController` class from `lib/buttons.dart` to a new file `lib/animation_controller.dart`.
  - Moves `_LetterDetailsPage` class from `lib/buttons.dart` to a new file `lib/letter_details_page.dart`.

* **Code Readability and Short Functions**:
  - Refactors `LargeLetterButton` class `build` method in `lib/buttons.dart` for readability.
  - Refactors `emojiDialog` method in `lib/emoji.dart` into smaller functions for readability.
  - Refactors `SettingsPageState` class methods in `lib/settings.dart` into smaller functions for readability.

* **Imports**:
  - Adds necessary imports in `lib/animation_controller.dart` and `lib/letter_details_page.dart`.
  - Updates imports in `lib/buttons.dart` to reflect the moved classes.
